+++ src/main/java/Config/Routes/POSTRoute.java
@@ -3,8 +3,8 @@ package Config.Routes;
 import Config.Method;
 import Controller.Handler.Handler;
 
-public class GetRoute extends Route {
-	public GetRoute(Method method, String path, Handler handler) {
+public class POSTRoute extends Route {
+	public POSTRoute(Method method, String path, Handler handler) {
 		this.method = method;
 		this.path = path;
 		this.handler = handler;
