+++ src/main/java/Config/Routes/RouteFactory.java
@@ -3,19 +3,17 @@ package Config.Routes;
 import Config.Method;
 import Controller.Handler.Handler;
 
-public class RouteFactory {
-	SimpleRouteFactory factory;
- 
-	public RouteFactory(SimpleRouteFactory factory) {
-		this.factory = factory;
-	}
- 
-	public Route createRoute(Method method, String path, Handler handler) {
-		Route route;
- 
-		route = factory.createRoute(method, path, handler);
+import static Config.Method.*;
 
+public class RouteFactory {
+
+	public Route createRoute(Method method, String path, Handler handler) {
+		Route route = null;
+		if (method == GET) {
+			route = new GETRoute(method, path, handler);
+		} else if (method == POST) {
+			route = new POSTRoute(method, path, handler);
+		}
 		return route;
 	}
-
 }
