+++ src/main/java/Validator/Validator.java
@@ -1,5 +1,9 @@
 package Validator;
 
+import java.io.IOException;
+import java.net.Socket;
+import java.net.SocketException;
+
 public class Validator {
 
   public static boolean validArgsLength(String[] args) {
@@ -16,6 +20,15 @@ public class Validator {
       return true;
     } else {
       System.err.println("Usage: StartServer [-p] [PORT number]");
+      return false;
+    }
+  }
+
+  public static boolean portIsNotAvailable(String port) throws IOException {
+    try {
+      (new Socket("127.0.0.1", Integer.parseInt(port))).close();
+      return true;
+    } catch (SocketException ignored) {
       return false;
     }
   }
